Add submodule install support Add gradle and some tmux todo files

+++ install/link.sh
@@ -6,8 +6,14 @@ echo "creating symlinks"
 linkables=$( find -H "$DOTFILES" -maxdepth 3 -name '*.symlink' )
 for file in $linkables ; do
     target="$HOME/.$( basename $file ".symlink" )"
-    echo "  creating symlink for $file"
-    # $file is already absolute no need to specify DOTFILES again
-    # ln -s $DOTFILES/$file $target
-    ln -s $file $target
+    echo -n "  creating symlink for $file"
+    if [ -L "$target" ]; then
+      echo " -- link exists"
+    elif [ -e "$target" ]; then
+      echo " -- OVERRIDDEN"
+    else
+      # $file is already absolute no need to specify DOTFILES again
+      # ln -s $DOTFILES/$file $target
+      ln -s $file $target
+    fi
 done
